build: using ‘config’ for ‘exposes’ config instead of ‘exports’
We intended to use “exports” to denote which ‘modules’ an app exposes for module federation.  Unfortunately, adding exports is meaningful to tsc and webpack, and changes how they expect to use the package in ways we don’t want.  Instead, I’m just putting “exposes” into the “config” block and passing that through to webpack so that we’re not using something intended for a different purpose.